1 version witohut fact

+++ fullcars-cliente/src/main/java/dtos/SalesData.java
@@ -1,19 +1,20 @@
 package dtos;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
 public class SalesData {
-    private long amount;
+    private BigDecimal amount;
     private LocalDate date;
 
-    public SalesData(long amount, LocalDate date) {
+    public SalesData(BigDecimal amount, LocalDate date) {
         this.amount = amount;
         this.date = date;
     }
     
-    public long getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public LocalDate getDate() { return date; }
 }
