+++ fullcars-cliente/src/main/java/model/client/entities/PurchaseDetail.java
@@ -1,5 +1,7 @@
 package model.client.entities;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,7 +14,7 @@ public class PurchaseDetail extends Detail{
 
     private Purchase purchase;
 
-	public PurchaseDetail(Integer quantity, Long unitPrice, CarPart product) {
+	public PurchaseDetail(Integer quantity, BigDecimal unitPrice, CarPart product) {
 		super(quantity, unitPrice, product);
 	}
 }
